refactor(agentic-ai): nest Anthropic effort/thinking enums in the request model, inline reasoning validation
AnthropicEffort and ThinkingMode lived in the provider package even though
they are wire-format dropdown enums owned by the request model, unlike
ThinkingDisplay which was already nested there; move them to match.
AnthropicReasoningValidator had a single caller and a three-line check, so
fold it into AnthropicMessageRequestConverter instead of keeping it as its
own class. ConversationSnapshot now exposes the leading system message
directly (windowing never evicts it), so the request converter no longer
re-scans the whole message list to extract it.
