+++ connectors/agentic-ai/connector-agentic-ai/src/main/java/io/camunda/connector/agenticai/aiagent/memory/ConversationSnapshot.java
@@ -7,9 +7,11 @@
 package io.camunda.connector.agenticai.aiagent.memory;
 
 import io.camunda.connector.agenticai.aiagent.model.message.Message;
+import io.camunda.connector.agenticai.aiagent.model.message.SystemMessage;
 import io.camunda.connector.agenticai.aiagent.model.tool.ToolDefinition;
 import java.util.List;
 import java.util.Objects;
+import org.jspecify.annotations.Nullable;
 
 /**
  * Immutable snapshot of the windowed conversation sent to the LLM. Carries the filtered message
@@ -22,5 +24,13 @@ public record ConversationSnapshot(List<Message> messages, List<ToolDefinition> 
     Objects.requireNonNull(toolDefinitions);
     messages = List.copyOf(messages);
     toolDefinitions = List.copyOf(toolDefinitions);
+  }
+
+  // Windowing never evicts the system message (see MessageWindowFilter), so it is always the
+  // leading message if present.
+  public @Nullable SystemMessage systemMessage() {
+    return !messages.isEmpty() && messages.getFirst() instanceof SystemMessage systemMessage
+        ? systemMessage
+        : null;
   }
 }
